blockchain-publisher/requirements.txt
-i https://pypi.org/simple
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
base58==2.0.1; python_version >= '3.5'
bitarray==1.2.2
certifi==2020.11.8
chardet==3.0.4
cytoolz==0.11.0; implementation_name == 'cpython'
eth-abi==2.1.1; python_version >= '3.6' and python_version < '4'
eth-account==0.5.4; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1; python_version >= '3.6' and python_version < '4'
eth-typing==2.2.2; python_version >= '3.5' and python_version < '4'
eth-utils==1.9.5; python_version >= '3.5' and python_full_version != '3.5.2' and python_version < '4'
hexbytes==0.2.1; python_version >= '3.6' and python_version < '4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ipfshttpclient==0.7.0a1; python_full_version >= '3.5.4' and python_full_version not in '3.6.0, 3.6.1, 3.7.0, 3.7.1'
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
netaddr==0.8.0
paho-mqtt==1.5.1
parsimonious==0.8.1
protobuf==3.13.0
py-solc-x==1.0.0
py-solc==3.2.0
pycryptodome==3.9.9
pyrsistent==0.17.3; python_version >= '3.5'
requests==2.24.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rlp==2.0.0
semantic-version==2.8.5; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
toolz==0.11.1; python_version >= '3.5'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
varint==1.0.2
web3==5.13.0
websockets==8.1; python_full_version >= '3.6.1'
